sops: activation: Avoid leaking the secret upon creation.
Secrets files were created with the access mode 644. This implies that
all secrets are readable by others until they are chowned again by
activate-secrets. This patch changes the way files are created to first
create a temporary file, setting the minimal set of necessary permissions
for the current user and writing the secret contents to the
file. Finally the temporary file is renamed to the actual secret name.

* modules/sops/activation.scm (activate-secrets): Avoid leaking secrets
to all other users.

Fixes: https://github.com/fishinthecalculator/sops-guix/issues/14